==> Data Preparation/02-Data match.ipynb <==
import geopandas as gpd
import pandas as pd
from shapely.geometry import Point
import transbigdata as tbd
import matplotlib.pyplot as plt
import numpy as np

districts = gpd.read_file('UrbanEVSupplemental\shenzhen_districts\Shenzhen.shp')
print(districts.columns)
print(districts.shape)
districts.head(2)

zone = pd.read_csv('zone-information.csv')
print(zone.columns)
print(zone.shape)
zone.head(2)

zone['TAZID'] = zone['TAZID'].astype(int)
districts['TAZID'] = districts['TAZID'].astype(int)
zone_geo = pd.merge(zone, districts, on='TAZID', how='left')

zone_geo = zone_geo[['TAZID', 'area', 'geometry']]
zone_geo = gpd.GeoDataFrame(zone_geo,geometry='geometry')
print(zone_geo.shape)
print(type(zone_geo))
zone_geo.head(2)

zone_geo = zone_geo.set_crs(epsg=3857)
zone_geo = zone_geo.to_crs(epsg=4326)
zone_geo.head(2)

## Landuse

landuse = gpd.read_file('guangdong_latest_shapefile/guangdong_latest_shapefile/gis_osm_landuse_a_free_1.shp')
print(landuse.columns)
print(landuse.shape)
landuse.head(2)

station = pd.read_csv('UrbanEVDataset/20220901-20230228_station-raw/station_information.csv')
station = station[['station_id', 'longitude', 'latitude', 'TAZID']]

#转换为GeoDataFrame
station['geometry'] = gpd.points_from_xy(station['longitude'],station['latitude'])
station = gpd.GeoDataFrame(station)
station = station.set_crs('EPSG:4326')
print(type(station))
print(station.shape)
station.head(2)

#第一步点和区域进行匹配
matched_all_point = gpd.sjoin(station, landuse, how='left', predicate='within')
matched_all_point = matched_all_point.drop_duplicates(subset=['station_id'])
matched_all_point = matched_all_point[['station_id', 'longitude', 'latitude', 'TAZID', 'fclass', 'geometry']]
matched_all_point.head(2)

matched_in_point = matched_all_point[matched_all_point.notna().all(axis=1)]
unmatched_point = matched_all_point[matched_all_point['fclass'].isna()]

matched_out_point = gpd.sjoin_nearest(unmatched_point, landuse, how='left')
matched_out_point = matched_out_point.drop_duplicates(subset=['station_id'])
matched_out_point = matched_out_point[['station_id', 'longitude', 'latitude', 'TAZID', 'fclass_right', 'geometry']]
matched_out_point.columns = ['station_id', 'longitude', 'latitude', 'TAZID', 'fclass', 'geometry']

matched_point = pd.concat([matched_in_point, matched_out_point], axis=0)
matched_point = matched_point.reset_index()
matched_point = matched_point.drop('index',axis=1)

print(matched_all_point.shape)
print(matched_in_point.shape)
print(matched_out_point.shape)
print(matched_point.shape)

matched_point.head(2)

from scipy.stats import entropy

def calculate_shannon_entropy(group):
    value_counts = group['fclass'].value_counts()
    proportions = value_counts / len(group)
    return entropy(proportions, base=np.e)  # 使用自然对数

landuse_shannon = matched_point.groupby('TAZID').apply(calculate_shannon_entropy).reset_index()
landuse_shannon.columns = ['TAZID', 'shannon']
landuse_shannon.head(2)

result_landuse = pd.merge(zone, landuse_shannon, on='TAZID', how='left')
result_landuse = result_landuse[['TAZID', 'shannon']]
result_landuse.head(2)

result_landuse.to_csv('landuse.csv',index=None)

## Road network

roads = gpd.read_file('guangdong_latest_shapefile/guangdong_latest_shapefile/gis_osm_roads_free_1.shp')
#是否需要挑选出特定类型的道路，人行道等是否需要加入进去
roads = roads[['osm_id', 'geometry']]
road = roads.set_crs(epsg=4326)
print(roads.shape)
roads.head(2)

zone_geo.head(2)

def calc_road_length_by_region(roads_gdf, regions_gdf, region_id_col, target_crs="EPSG:32650"):
    # 转换到目标投影坐标系（米）
    roads_proj = roads_gdf.to_crs(target_crs)
    regions_proj = regions_gdf.to_crs(target_crs)

    # 计算交集
    inter = gpd.overlay(roads_proj, regions_proj, how="intersection")

    # 计算交集部分的长度（米）
    inter["length"] = inter.geometry.length

    # 按行政区聚合
    result = inter.groupby(region_id_col)["length"].sum().reset_index()

    # 合并回原始 regions（保留经纬度坐标系）
    regions_with_len = regions_gdf.merge(result, on=region_id_col, how="left")

    return regions_with_len

roads_matched = calc_road_length_by_region(roads, zone_geo, "TAZID")
roads_matched['road_density'] = (roads_matched['length']/roads_matched['area'])*1e3
roads_matched.head(2)

roads_matched[['TAZID','road_density']].to_csv('road_density.csv',index=None)
