==> Data Preparation/09-GDP.ipynb <==
import geopandas as gpd
import numpy as np
import pandas as pd
import rasterio
import rasterio.mask
import matplotlib.pyplot as plt

zone_information = pd.read_csv("/gpfsnyu/home/yw9871/New data/Original_data/zone-information.csv")
#test = zone_information.drop_duplicates(subset=['district'])
zone_information.head(2)

gdp = pd.read_excel("/gpfsnyu/home/yw9871/New data/Original_data/GDP.xlsx")
gdp.head(2)

gdp['weight2022'] = gdp['GDP2022'] / gdp['GDP2022'].sum()
gdp['weight2023'] = gdp['GDP2023'] / gdp['GDP2023'].sum()
gdp['gdp_weight'] = (gdp['weight2022'] + gdp['weight2023']) / 2
gdp.head(2)

ev_volume = pd.read_excel("/gpfsnyu/home/yw9871/New data/Original_data/EV volume.xlsx")
ev_volume.columns = ['time','city','district','ev_volume']
ev_volume['volume_weight'] = ev_volume['ev_volume'] / ev_volume['ev_volume'].sum()
ev_volume.head(2)

district = pd.merge(gdp, ev_volume, on='district', how='left')
district = district[['district', 'gdp_weight', 'volume_weight']]
district.head(2)

zone = pd.merge(zone_information, district, on='district', how='left')
zone = zone[['TAZID', 'gdp_weight', 'volume_weight']]
zone.head(2)

zone.to_csv("/gpfsnyu/home/yw9871/New data/Results/gdp_volume.csv",index=None)
